% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMoperational.R
\name{MMoperational}
\alias{MMoperational}
\title{Crea matriz multidimensional de archivos operacionales del modelo CFSv2}
\usage{
MMoperational(variable = "prate", direct)
}
\value{
Matriz multidimensional, donde cada elemento es un dataframe asociado a una coordenada.
}
\description{
Información mas detallada.
}
