% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMreforecast.R
\name{MMreforecast}
\alias{MMreforecast}
\title{Crea matriz multidimensional de archivos retrospectivos del modelo CFSv2}
\usage{
MMreforecast(variable = "prate", direct)
}
\value{
Matriz multidimensional, donde cada elemento es un dataframe asociado a una coordenada.
}
\description{
Información mas detallada.
}
