% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelcfs.R
\name{modelcfs}
\alias{modelcfs}
\title{Crea serie predictora para período operacional, desagregada espacialmente y con el sesgo corregido.}
\usage{
modelcfs(tol = 0.5, n = 12, p = 0.3, coef, MMoperational, auxfiles)
}
\value{
Matriz multidimensional
}
\description{
Información mas detallada.
}
