% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelcfsrealtime.R
\name{modelcfsrealtime}
\alias{modelcfsrealtime}
\title{Crea serie predictora en tiempo real, desagregada espacialmente y con el sesgo corregido.}
\usage{
modelcfsrealtime(variable, tol = 0.5, p = 0.3, n = 12, coef, direct,
  auxfiles)
}
\value{
Matriz multidimensional
}
\description{
Información mas detallada.
}
