% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Realiza los pre-procesos de descarga, corte y conversión la información del modelo CFSv2 operacional y retrospectivo.}
\usage{
preprocess(product, variable, start, end, lat1, lon1, lat2, lon2, direct)
}
\value{
Archivos NetCDF para el rango de tiempocortados en la zona de estudio
}
\description{
Información mas detallada.
}
