% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocesscfsrealtime.R
\name{preprocesscfsrealtime}
\alias{preprocesscfsrealtime}
\title{Realiza los pre-procesos de descarga, corte y conversión la información del modelo CFSv2 en tiempo real}
\usage{
preprocesscfsrealtime(variable, lat1, lat2, lon1, lon2, direct)
}
\value{
Archivos NetCDF para el rango de tiempocortados en la zona de estudio
}
\description{
Información mas detallada.
}
