% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDBCcfs.R
\name{SDBCcfs}
\alias{SDBCcfs}
\title{Encuentra una matriz de coeficientes de ajuste gamma para la SDBC, por cada estación meteorológica y por cada mes.}
\usage{
SDBCcfs(variable, MMreforecast, auxfiles, tol = 0.5)
}
\value{
Matriz multidimensional
}
\description{
Información mas detallada.
}
