% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verificationcfs.R
\name{verificationcfs}
\alias{verificationcfs}
\title{Encuentra matriz de estadígrafos de bondad de ajuste, para cada estación meteorológica y por cada mes.}
\usage{
verificationcfs(variable = "prate", coef, auxfiles, modelcfsoperational)
}
\value{
Dataframe
}
\description{
Información mas detallada.
}
